requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.2
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.5
openai==1.75.0
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-engineio==4.12.0
python-http-client==3.3.7
python-socketio==5.13.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
sendgrid==6.11.0
simple-websocket==1.1.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
sympy==1.13.3
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
